requirements.txt
pyyaml
pymongo
requests
pytest
ruamel.yaml
psycopg2